requirements.txt
beautifulsoup4==4.6.3
billboard.py==5.0.0
